requirements.txt
cryptography==41.0.7
opencv-python==4.9.0.80
Pillow==10.2.0
numpy==1.26.4